refactor(cli): move Android bin resolvers to core
emulatorBin and adbBin are pure path builders, and three more call sites
need them: emulatorSetup, the doctor SDK checks, and the AVD installer.
