+++ src/shell/appium/createAndroidEmulator.test.ts
@@ -1,12 +1,6 @@
 import { afterEach, describe, expect, it, mock } from "bun:test";
-import { join } from "node:path";
-
 import type { AdbFn, SpawnFn } from "./createAndroidEmulator.js";
-import {
-  adbBin,
-  createAndroidEmulator,
-  emulatorBin,
-} from "./createAndroidEmulator.js";
+import { createAndroidEmulator } from "./createAndroidEmulator.js";
 
 afterEach(() => {
   mock.restore();
@@ -141,44 +135,5 @@ describe("createAndroidEmulator", () => {
     // Wait for the background wait-for-device to reject; an unhandled rejection here
     // would crash the Bun process and fail the test suite.
     await new Promise<void>((r) => setTimeout(r, 100));
-  });
-});
-
-describe("emulatorBin", () => {
-  const home = join("/opt", "android-sdk");
-
-  it("returns the extension-less path on linux and macOS", () => {
-    expect(emulatorBin(home, "linux")).toBe(join(home, "emulator", "emulator"));
-    expect(emulatorBin(home, "darwin")).toBe(
-      join(home, "emulator", "emulator"),
-    );
-  });
-
-  it("returns emulator.exe on win32", () => {
-    expect(emulatorBin(home, "win32")).toBe(
-      join(home, "emulator", "emulator.exe"),
-    );
-  });
-
-  it("falls back to the bare name on PATH when ANDROID_HOME is unset", () => {
-    expect(emulatorBin(undefined, "linux")).toBe("emulator");
-    expect(emulatorBin(undefined, "win32")).toBe("emulator.exe");
-  });
-});
-
-describe("adbBin", () => {
-  const home = join("/opt", "android-sdk");
-
-  it("returns the extension-less path on linux and macOS", () => {
-    expect(adbBin(home, "linux")).toBe(join(home, "platform-tools", "adb"));
-  });
-
-  it("returns adb.exe on win32", () => {
-    expect(adbBin(home, "win32")).toBe(join(home, "platform-tools", "adb.exe"));
-  });
-
-  it("falls back to the bare name on PATH when ANDROID_HOME is unset", () => {
-    expect(adbBin(undefined, "linux")).toBe("adb");
-    expect(adbBin(undefined, "win32")).toBe("adb.exe");
   });
 });
