+++ src/shell/appium/createAndroidEmulator.ts
@@ -1,6 +1,7 @@
 import { execFile, spawn } from "node:child_process";
-import { join } from "node:path";
 import { promisify } from "node:util";
+
+import { adbBin, emulatorBin } from "~/core/androidBins.js";
 
 const execFileAsync = promisify(execFile);
 const defaultBootTimeoutMs = 120_000;
@@ -8,28 +9,6 @@ const pollIntervalMs = 2_000;
 
 function androidHome(): string | undefined {
   return process.env["ANDROID_HOME"] ?? process.env["ANDROID_SDK_ROOT"];
-}
-
-// The SDK ships emulator.exe and adb.exe on Windows. libuv appends .exe only
-// during a PATH search, so an explicit ANDROID_HOME path needs the extension.
-function withExeSuffix(name: string, platform: NodeJS.Platform): string {
-  return platform === "win32" ? `${name}.exe` : name;
-}
-
-export function emulatorBin(
-  home: string | undefined,
-  platform: NodeJS.Platform,
-): string {
-  const name = withExeSuffix("emulator", platform);
-  return home ? join(home, "emulator", name) : name;
-}
-
-export function adbBin(
-  home: string | undefined,
-  platform: NodeJS.Platform,
-): string {
-  const name = withExeSuffix("adb", platform);
-  return home ? join(home, "platform-tools", name) : name;
 }
 
 export type SpawnFn = (bin: string, args: string[]) => { stop: () => void };
